feat: add schema

+++ src/services/supabase/TransactionsService.ts
@@ -1,46 +1,34 @@
 import { supabase } from "@/helpers/supabase";
-import { NewTransaction } from "@/interfaces/Transactions";
+import { Transaction } from "@/schema";
 
 const TransactionsService = () => {
-  async function addTransaction({user_id, flow, amount, wallet, trx_id, message }: NewTransaction) {
+  async function addTransaction({ user_id, flow, amount, wallet, trx_id, message }: Transaction) {
     const { error } = await supabase
-      .from('transactions')
-      .insert({user_id, flow, amount, wallet, trx_id, message });
-      if (error) throw error;
+      .from("transactions")
+      .insert({ user_id, flow, amount, wallet, trx_id, message });
+    if (error) throw error;
   }
   async function getTransactionById(trxId: string) {
-    const { data, error } = await supabase
-      .from('transactions')
-      .select("*")
-      .eq('trx_id', trxId); 
-  
+    const { data, error } = await supabase.from("transactions").select("*").eq("trx_id", trxId);
+
     if (error) throw error;
     return data;
   }
   async function getTransactionsByUserId(user_id: string) {
-    const { data, error } = await supabase
-      .from('transactions')
-      .select("*")
-      .eq('user_id', user_id); 
-  
+    const { data, error } = await supabase.from("transactions").select("*").eq("user_id", user_id);
+
     if (error) throw error;
     return data;
   }
   async function deleteTransactionsByUserId(userId: string) {
-    const { error } = await supabase
-      .from('transactions')
-      .delete()
-      .eq('user_id', userId);
-    
+    const { error } = await supabase.from("transactions").delete().eq("user_id", userId);
+
     if (error) throw error;
     return;
   }
   async function deleteTransactionById(trxId: string) {
-    const { error } = await supabase
-      .from('transactions')
-      .delete()
-      .eq('trx_id', trxId);
-    
+    const { error } = await supabase.from("transactions").delete().eq("trx_id", trxId);
+
     if (error) throw error;
     return;
   }
