+++ src/services/supabase/UsersServices.ts
@@ -1,59 +1,50 @@
 import { supabase } from "@/helpers/supabase";
-import {user as admin} from '@/composables/useAuthUser';
-import { NewUser } from "@/interfaces/Users";
+import { useAuthUser } from "@/composables/useAuthUser";
+import { User } from "@/schema";
+
+const { user: admin } = useAuthUser();
 
 const UsersService = () => {
-  async function addUser({admin_id, user_id, name, username, color_profile}: NewUser) {
-    const { error } = await supabase
-    .from('users')
-    .insert({
+  async function addUser({ admin_id, user_id, name, username, color_profile }: User) {
+    const { error } = await supabase.from("users").insert({
       admin_id,
       user_id,
       name,
       username,
-      color_profile,    
-    });    
+      color_profile,
+    });
     if (error) throw error;
   }
   async function getUsers() {
-    const { data, error } = await supabase
-    .from('users')
-    .select()
-    .eq('admin_id',admin?.value.id);
+    const { data, error } = await supabase.from("users").select().eq("admin_id", admin?.value.id);
     if (error) throw error;
     return data;
   }
   async function getUserById(userId: string) {
-    const { data, error } = await supabase
-      .from('users')
-      .select("*")
-      .eq('user_id', userId);
+    const { data, error } = await supabase.from("users").select("*").eq("user_id", userId);
     if (error) throw error;
     return data;
   }
-  async function updateUser(userId: string, name : string, username: string, color_profile: string) {
+  async function updateUser(userId: string, name: string, username: string, color_profile: string) {
     const { error } = await supabase
-      .from('users')
+      .from("users")
       .update({ name, username, color_profile })
-      .eq('user_id', userId);
-      if (error) throw error;
+      .eq("user_id", userId);
+    if (error) throw error;
   }
-  async function updateName(currentName: string,  userId: string){
+  async function updateName(currentName: string, userId: string) {
     const { error } = await supabase
-      .from('users')
+      .from("users")
       .update({ name: currentName })
-      .eq('user_id', userId);
-      if (error) throw error;
+      .eq("user_id", userId);
+    if (error) throw error;
   }
   async function deleteUserByUserId(userId: string) {
-    const { error } = await supabase
-      .from('users')
-      .delete()
-      .eq('user_id', userId);
-      if (error) throw error;
+    const { error } = await supabase.from("users").delete().eq("user_id", userId);
+    if (error) throw error;
     return;
   }
-  
+
   return {
     addUser,
     getUsers,
